+++ docs/internal/app/guides/index.ts
@@ -1,3 +1,5 @@
 export * from './home.component';
 export * from './introduction/getting-started.component';
 export * from './core-concepts/layouts.component';
+export * from './core-concepts/components.component';
+export * from './core-concepts/themeing.component';