+++ packages/components/CHANGES.md
@@ -1,5 +1,13 @@
 # Roxie Components
 
+## [1.2.0]
+> Sep 26, 2019
+
+* Implement `Container` component
+* Implement `Dialog` component
+* Implement `Chart` component
+* Clean up code
+
 ## [0.9.4]
 > Apr 6, 2019